feat(ast): remove serde skip for symbol_id and reference_id
