+++ crates/oxc_syntax/Cargo.toml
@@ -34,5 +34,5 @@ tsify        = { workspace = true, optional = true }
 wasm-bindgen = { workspace = true, optional = true }
 [features]
 default = []
-serde   = ["dep:serde", "bitflags/serde"]
+serde   = ["dep:serde", "bitflags/serde", "oxc_index/serde"]
 wasm    = ["dep:tsify", "dep:wasm-bindgen"]
