+++ crates/oxc_syntax/src/lib.rs
@@ -4,6 +4,7 @@ pub mod assumptions;
 pub mod class;
 pub mod identifier;
 pub mod module_record;
+pub mod node;
 pub mod operator;
 pub mod precedence;
 pub mod reference;
